work on feedback from kyle
